requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.1
distro==1.9.0
Django==5.1.1
djangorestframework==3.15.2
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==2.1.1
openai==1.44.0
packaging==24.1
pydantic==2.9.0
pydantic_core==2.23.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
setuptools==74.1.2
sniffio==1.3.1
sqlparse==0.5.1
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.10.0
